Remove workerInternalBlockWorker
